recognition_platform/requirements/base.txt
Django==3.2.3
django-environ==0.4.5
djangorestframework==3.12.4
gunicorn==20.0.4
whitenoise==5.2.0
psycopg2-binary==2.8.6
Pillow
factory-boy==3.2.0
Faker==8.10.1
pytest-django==4.4.0
flake8==3.9.2
django-cors-headers==3.7.0
cryptography==2.3.1
drf-jwt==1.13.3
pyjwt==1.7.1
requests==2.22.0
django-filter==2.4.0
